feat: rejigs clamav stuff to use clamscan instead of clamdscan
We're using a systemd timer for scheduling, and a systemd service to run
a custom script, av-scan, that logs to /var/log/clamav.log, and writes
the summary of the last run to ~/last-clamscan.log.

The plan is to show the contents of that file whenever a new terminal is
launched, *and* to use the signal library to send a notification when
malware is detected.